machine_learning/requirements.txt
pillow>=5.4.1
matplotlib>=3.0.3
tensorboardX>=1.6
torch>=1.0.0
torchvision>=0.2.1
tqdm>=4.31.1